% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_codelgt.R
\docType{data}
\name{df_codelgt}
\alias{df_codelgt}
\title{Table de liaison lgt type cd et libelles
Table de liaison entre les codes des types de logements et leurs libelles}
\format{
a dataframe with 4 rows and two variables:
\describe{
\item{variable}{le code de la variable}
\item{libelle}{le libelle de la variable en facteur}
}
}
\usage{
df_codelgt
}
\description{
Table de liaison lgt type cd et libelles
Table de liaison entre les codes des types de logements et leurs libelles
}
\keyword{datasets}
