% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_moislib.R
\docType{data}
\name{df_moislib}
\alias{df_moislib}
\title{Table de liaison mois chiffre et libelle
Table de liaison mois chiffre et libelle}
\format{
a dataframe with 372 rows and 7 variables:
\describe{
\item{date}{date au format annee mois 200001}
\item{mois}{mois au format chiffre 01}
\item{moisc}{mois au format janv.}
\item{moiscl}{mois au format janvier}
\item{moislib}{date au format mois annee janv. 00 }
\item{moislibl}{date au format mois annee janv. 2000 }
}
}
\usage{
df_moislib
}
\description{
Table de liaison mois chiffre et libelle
Table de liaison mois chiffre et libelle
}
\keyword{datasets}
