% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn00_verifie_existe_modele.R
\name{fn00_verifie_existe_modele}
\alias{fn00_verifie_existe_modele}
\title{fn00_verifie_existe_modele}
\usage{
fn00_verifie_existe_modele(x = params)
}
\arguments{
\item{x}{character oui ou non}
}
\value{
chemin vers le modele
}
\description{
Verifie que si modele_plaquette est oui le fichier sla est bien dans 3_tables
}
