% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn01_cree_repertoires.R
\name{fn01_cree_repertoires}
\alias{fn01_cree_repertoires}
\title{fn01_cree_repertoires
on cree les repertoires 2_data, 4_resultats dans le nouveau dossier s ils n existent pas.}
\usage{
fn01_cree_repertoires(dossier = here::here())
}
\arguments{
\item{dossier}{le repertoire ou est le projet}
}
\value{
nothing
}
\description{
fn01_cree_repertoires
on cree les repertoires 2_data, 4_resultats dans le nouveau dossier s ils n existent pas.
}
