% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn03_lit_tab_excel.R
\name{fn03_lit_tab_excel}
\alias{fn03_lit_tab_excel}
\title{fn03_lit_tab_excel}
\usage{
fn03_lit_tab_excel(x = params$annee_mois)
}
\arguments{
\item{x}{vecteur caractere path to file}
}
\value{
liste
}
\description{
Lit le tableau excel du SDES
}
