% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn04_modifie1_liste.R
\name{fn04_modifie1_liste}
\alias{fn04_modifie1_liste}
\title{fn04_modifie1_liste}
\usage{
fn04_modifie1_liste(data = lsm)
}
\arguments{
\item{data}{liste}
}
\value{
liste
}
\description{
modifie liste valeurs en liste valeurs moyennes non arrondies.
}
