% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn05_modifie2_liste.R
\name{fn05_modifie2_liste}
\alias{fn05_modifie2_liste}
\title{fn05_modifie2_liste}
\usage{
fn05_modifie2_liste(data = lsm_12m0)
}
\arguments{
\item{data}{liste}
}
\value{
liste
}
\description{
modifie liste valeurs moyennes en liste valeurs moyennes arrondies.
}
