% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn06_graphe_mensuel_np.R
\name{fn06_graphe_mensuel_np}
\alias{fn06_graphe_mensuel_np}
\title{fn06_graphe_mensuel_np}
\usage{
fn06_graphe_mensuel_np(x = "aut")
}
\arguments{
\item{x}{charactere aut ou com}
}
\value{
graphe
}
\description{
Affichage du graphe des logements autorises par mois
}
