% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn07_graphe_lgt_type_evol.R
\name{fn07_graphe_lgt_type_evol}
\alias{fn07_graphe_lgt_type_evol}
\title{fn07_graphe_lgt_type_evol}
\usage{
fn07_graphe_lgt_type_evol(x = "aut")
}
\arguments{
\item{x}{caractere aut ou com}
}
\value{
graphe
}
\description{
fn07_graphe_lgt_type_evol
}
