% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn07_graphe_lgt_type_evol_area.R
\name{fn07_graphe_lgt_type_evol_area}
\alias{fn07_graphe_lgt_type_evol_area}
\title{fn07_graphe_lgt_type_evol_area}
\usage{
fn07_graphe_lgt_type_evol_area(x = "aut")
}
\arguments{
\item{x}{character aut ou com}
}
\value{
graphe
}
\description{
Affiche un graphe surfacique
}
