% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn08_tableau_bilan.R
\name{fn08_tableau_bilan}
\alias{fn08_tableau_bilan}
\title{fn08_tableau_bilan}
\usage{
fn08_tableau_bilan(x = "aut")
}
\arguments{
\item{x}{charactere aut ou com}
}
\value{
dataframe
}
\description{
fn08_tableau_bilan
}
