% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn09_tableau_lgt_tous_territoire.R
\name{fn09_tableau_lgt_tous_territoire}
\alias{fn09_tableau_lgt_tous_territoire}
\title{fn09_tableau_lgt_tous_territoire}
\usage{
fn09_tableau_lgt_tous_territoire(x = "aut")
}
\arguments{
\item{x}{caractere aut ou com}
}
\value{
dataframe
}
\description{
Creation des tableaux au format texte pour la plaquette
}
