% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn10_tableau_lgt_type.R
\name{fn10_tableau_lgt_type}
\alias{fn10_tableau_lgt_type}
\title{fn10_tableau_lgt_type}
\usage{
fn10_tableau_lgt_type(x = "toto")
}
\arguments{
\item{x}{caractere aut ou com}
}
\value{
dataframe
}
\description{
Creation des tableaux au format texte pour la plaquette
}
