% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn11_evol1an_type_lgt_bonus1.R
\name{fn11_evol1an_type_lgt_bonus1}
\alias{fn11_evol1an_type_lgt_bonus1}
\title{fn11_evol1an_type_lgt_bonus1}
\usage{
fn11_evol1an_type_lgt_bonus1(x = "aut")
}
\arguments{
\item{x}{caractere aut ou com}
}
\value{
liste
}
\description{
Analyse evolution sur un an par type de logement
}
