% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn11_graphe_barres_lgt_type_territoire.R
\name{fn11_graphe_barres_lgt_type_territoire}
\alias{fn11_graphe_barres_lgt_type_territoire}
\title{fn11_graphe_barres_lgt_type_territoire}
\usage{
fn11_graphe_barres_lgt_type_territoire(x = "aut")
}
\arguments{
\item{x}{caractere aut ou com}
}
\value{
graphique
}
\description{
Graphiques à barre, composition des nouveaux logements par typologie
par territoire. Voir page 2 et 3
}
