% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn12_graphe_lgt_tous_base100.R
\name{fn12_graphe_lgt_tous_base100}
\alias{fn12_graphe_lgt_tous_base100}
\title{fn12_graphe_lgt_tous_base100}
\usage{
fn12_graphe_lgt_tous_base100(x = "aut")
}
\arguments{
\item{x}{caractere aut ou com}
}
\value{
graphe
}
\description{
fn12_graphe_lgt_tous_base100
}
