% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn13_graphe_lgt_tous_autcom_dep.R
\name{fn13_graphe_lgt_tous_autcom_dep}
\alias{fn13_graphe_lgt_tous_autcom_dep}
\title{fn13_graphe_lgt_tous_autcom_dep}
\usage{
fn13_graphe_lgt_tous_autcom_dep(data = lsm_12m0$DPT)
}
\arguments{
\item{data}{liste}
}
\value{
nothing
}
\description{
fn13_graphe_lgt_tous_autcom_dep
}
