% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn14_carte_region_lgt_tous_evolution.R
\name{fn14_carte_region_lgt_tous_evolution}
\alias{fn14_carte_region_lgt_tous_evolution}
\title{fn14_carte_region_lgt_tous_evolution}
\usage{
fn14_carte_region_lgt_tous_evolution(x = "aut")
}
\arguments{
\item{x}{caractere aut ou com}
}
\value{
graphique
}
\description{
fn14_carte_region_lgt_tous_evolution
}
