% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn15_enregistre_tab_4serveur.R
\name{fn15_enregistre_tab_4serveur}
\alias{fn15_enregistre_tab_4serveur}
\title{fn15_enregistre_tab_4serveur}
\usage{
fn15_enregistre_tab_4serveur(x = "aut")
}
\arguments{
\item{x}{caractere aut ou com}
}
\value{
nothing
}
\description{
Enregistrement du tableau pour saisie dans l'outil indicateur
}
