% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn16_lit_le_modele_sla.R
\name{fn16_lit_le_modele_sla}
\alias{fn16_lit_le_modele_sla}
\title{fn16_lit_le_modele_sla}
\usage{
fn16_lit_le_modele_sla(x = "txt_plaquette")
}
\arguments{
\item{x}{date annee etude}
}
\value{
liste
}
\description{
Scribus : texte, lit la plaquette modele, supprime les lignes et en cree
autant de vides que dans le nouveau texte des tableaux.
}
