% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn17_modifie_objet_by_id_newvalue.R
\name{fn17_modifie_objet_by_id_newvalue}
\alias{fn17_modifie_objet_by_id_newvalue}
\title{fn17_modifie_objet_by_id_newvalue}
\usage{
fn17_modifie_objet_by_id_newvalue(
  id_objet = ls_modele$t_images$id_objet,
  attr = ls_modele$t_images$attr,
  new_value = ls_modele$t_images$new_value
)
}
\arguments{
\item{id_objet}{identifiant scribus de l objet}

\item{attr}{le nom de l attribut}

\item{new_value}{la nouvelle valeur}
}
\value{
nothing
}
\description{
Scribus : Modifie la valeur d attribut d un node
}
