% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn18_maj_tableaux.R
\name{fn18_maj_tableaux}
\alias{fn18_maj_tableaux}
\title{fn18_maj_tableaux}
\usage{
fn18_maj_tableaux(data = ls_tab)
}
\arguments{
\item{data}{liste}
}
\value{
df
}
\description{
Scribus : tableaux, tableau des cellules a remplacer
}
