% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn18_scribus_actualise_tableau.R
\name{fn18_scribus_actualise_tableau}
\alias{fn18_scribus_actualise_tableau}
\title{fn18_scribus_actualise_tableau}
\usage{
fn18_scribus_actualise_tableau(data = ls_tab)
}
\arguments{
\item{data}{liste}
}
\value{
df
}
\description{
Scribus : tableaux, tableau des cellules a remplacer
}
