% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn18_tableaux_cellule_a_remplacer.R
\name{fn18_tableaux_cellule_a_remplacer}
\alias{fn18_tableaux_cellule_a_remplacer}
\title{fn18_tableaux_cellule_a_remplacer}
\usage{
fn18_tableaux_cellule_a_remplacer(data = ls_tab)
}
\arguments{
\item{data}{liste}
}
\value{
df
}
\description{
Scribus : tableaux, tableau des cellules a remplacer
}
