% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn19_scribus_modifie_tableau.R
\name{fn19_scribus_modifie_tableau}
\alias{fn19_scribus_modifie_tableau}
\title{fn19_scribus_modifie_tableau}
\usage{
fn19_scribus_modifie_tableau(x = "211587152", y = 8, z = "Corse-du-Sud")
}
\arguments{
\item{x}{id_objet caractere}

\item{y}{num_cell numerique}

\item{z}{texte}
}
\value{
nothing
}
\description{
scribus, Tableaux, ajoute les valeurs issues des calculs dans les tableaux
}
