% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn21_tableaux_met_en_forme_2ndeligne.R
\name{fn21_tableaux_met_en_forme_2ndeligne}
\alias{fn21_tableaux_met_en_forme_2ndeligne}
\title{fn21_tableaux_met_en_forme_2ndeligne}
\usage{
fn21_tableaux_met_en_forme_2ndeligne(x = toto)
}
\arguments{
\item{x}{df}
}
\value{
rien
}
\description{
Scribus, tableaux, mise en forme des sous-indicateurs Enregistre aussi toutes les modifications en sortie
}
