% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_utilitaire_importScribus.R
\name{fn99_utilitaire_importScribus}
\alias{fn99_utilitaire_importScribus}
\title{utilitaire_importScribus}
\usage{
fn99_utilitaire_importScribus(x)
}
\arguments{
\item{x}{le chemin d'acces au modele de plaquette}
}
\value{
nothing
}
\description{
enregistre le modele scribus dans data
}
