% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_rc_po_no_td_allcell.R
\name{fn_rc_po_no_td_allcell}
\alias{fn_rc_po_no_td_allcell}
\title{Scribus interroge le fichier sla pour extraire tous les nodes des cellules
d'un tableau.}
\usage{
fn_rc_po_no_td_allcell(x = "211587152")
}
\arguments{
\item{x}{id_objet numero de l objet}
}
\value{
nodes avec uniquement les cellules du tableau
}
\description{
Scribus interroge le fichier sla pour extraire tous les nodes des cellules
d'un tableau.
}
