% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_rc_po_no_td_numcell.R
\name{fn_rc_po_no_td_numcell}
\alias{fn_rc_po_no_td_numcell}
\title{Scribus interroge le fichier sla pour extraire tous les nodes d'une cellule
d'un tableau.}
\usage{
fn_rc_po_no_td_numcell(x = "211587152", y = 2)
}
\arguments{
\item{x}{num_objet id du tableau}

\item{y}{num_cell id de la cellule}
}
\value{
nodes de la cellule
}
\description{
Scribus interroge le fichier sla pour extraire tous les nodes d'une cellule
d'un tableau.
}
