% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_scri_ls_nodes_cellules_tableau.R
\name{fn_scri_ls_nodes_cellules_tableau}
\alias{fn_scri_ls_nodes_cellules_tableau}
\title{fn_scri_ls_nodes_cellules_tableau}
\usage{
fn_scri_ls_nodes_cellules_tableau(id_objet = "211587152", num_cell = 2)
}
\arguments{
\item{id_objet}{id unique du tableau}

\item{num_cell}{id de la cellule}
}
\value{
nodes de la cellules
}
\description{
Scribus interroge le fichier sla pour extraire le node de la cellule specifiee
d un tableau.
}
