% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_scri_ls_nodes_tableau.R
\name{fn_scri_ls_nodes_tableau}
\alias{fn_scri_ls_nodes_tableau}
\title{fn_scri_ls_nodes_tableau}
\usage{
fn_scri_ls_nodes_tableau(id_objet = "211587152")
}
\arguments{
\item{id_objet}{numero unique scribus du tableau}
}
\value{
nodes avec uniquement les cellules du tableau
}
\description{
Scribus interroge le fichier sla pour extraire tous les nodes des cellules
d un tableau.
}
