% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_regwkt.R
\docType{data}
\name{map_regwkt}
\alias{map_regwkt}
\title{Couche simplifiee des regions metropolitaines
Couche simplifiee des regions metropolitaines au format wkt.}
\format{
a dataframe with 13 rows and 5 variables:
\describe{
\item{id}{identifiant fictif}
\item{INSEE_REG}{les codes insee des regions au format caractere}
\item{X}{coord abcisse en lambert93}
\item{Y}{coord ordonnee en lambert93}
\item{geom_wkt}{les coordonnees des polygones en wkt}
}
}
\usage{
map_regwkt
}
\description{
Couche simplifiee des regions metropolitaines
Couche simplifiee des regions metropolitaines au format wkt.
}
\keyword{datasets}
