% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plaquette_sitadel.R
\docType{data}
\name{plaquette_sitadel}
\alias{plaquette_sitadel}
\title{Modele de plaquette sitadel
modele realise sous scribus 1.5.6.1}
\format{
texte
}
\usage{
plaquette_sitadel
}
\description{
Modele de plaquette sitadel
modele realise sous scribus 1.5.6.1
}
\keyword{datasets}
