% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uti_calcul_arrondi_moyenne_12mois.R
\name{uti_calcul_arrondi_moyenne_12mois}
\alias{uti_calcul_arrondi_moyenne_12mois}
\title{uti_calcul_arrondi_moyenne_12mois}
\usage{
uti_calcul_arrondi_moyenne_12mois(data)
}
\arguments{
\item{data}{df}
}
\value{
df
}
\description{
uti_calcul_arrondi_moyenne_12mois
}
