% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uti_transpose_liste.R
\name{uti_transpose_liste}
\alias{uti_transpose_liste}
\title{uti_transpose_liste}
\usage{
uti_transpose_liste(data = lsm_12m)
}
\arguments{
\item{data}{liste}
}
\value{
liste
}
\description{
uti_transpose_liste
}
