% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uti_txdiff_ligne.R
\name{uti_txdiff_ligne}
\alias{uti_txdiff_ligne}
\title{uti_txdiff_ligne}
\usage{
uti_txdiff_ligne(x = c(1, 2, 3))
}
\arguments{
\item{x}{un vecteur numerique}
}
\value{
un vecteur numerique
}
\description{
uti_txdiff_ligne
}
\examples{
uti_txdiff_ligne(c(1,2,3))
}
