% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_utilitaire_importFrom.R
\name{x_utilitaire_importFrom}
\alias{x_utilitaire_importFrom}
\title{utilitaire_importFrom}
\usage{
x_utilitaire_importFrom(x = "fn06_met_en_forme_liste")
}
\arguments{
\item{x}{nom d'un fichier de fonction}
}
\value{
les packages et fonctions utilisées
}
\description{
utile pour lister les fonctions externes utilisees par chaque fonction
}
