% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn02_verifie_mois_valide.R
\name{fn02_verifie_mois_valide}
\alias{fn02_verifie_mois_valide}
\title{fn02_verifie_mois_valide}
\usage{
fn02_verifie_mois_valide(x = params$annee_mois)
}
\arguments{
\item{x}{caractere au format moisannee ex 202202}
}
\value{
nothing
}
\description{
Verifie que la date entree dans les parametres correspond bien a un tableau present dans 2_data
}
