% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn11_evol1an_type_lgt_bonus2.R
\name{fn11_evol1an_type_lgt_bonus2}
\alias{fn11_evol1an_type_lgt_bonus2}
\title{fn11_evol1an_type_lgt_bonus2}
\usage{
fn11_evol1an_type_lgt_bonus2(x = "aut")
}
\arguments{
\item{x}{caractere aut ou com}
}
\value{
liste
}
\description{
Analyse evolution sur un an pour tous les logements
}
