% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn17_modifie_objet_by_id_newvalue.R
\name{fn17_modifie_objet_by_id_newvalue}
\alias{fn17_modifie_objet_by_id_newvalue}
\title{fn17_modifie_objet_by_id_newvalue}
\usage{
fn17_modifie_objet_by_id_newvalue(
  x = ls_modele$t_images$id_objet,
  y = ls_modele$t_images$attr,
  z = ls_modele$t_images$new_value
)
}
\arguments{
\item{x}{identifiant scribus de l objet}

\item{y}{le nom de l attribut}

\item{z}{la nouvelle valeur}
}
\value{
nothing
}
\description{
Scribus : Modifie la valeur d attribut d un node
}
