% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uti_calcul_arrondi_moyenne_12mois.R
\name{uti_calcul_arrondi_moyenne_12mois}
\alias{uti_calcul_arrondi_moyenne_12mois}
\title{uti_calcul_arrondi_moyenne_12mois}
\usage{
uti_calcul_arrondi_moyenne_12mois(data)
}
\arguments{
\item{data}{df}
}
\value{
df
}
\description{
Compare la valeur arrondie a 100 du tous logements a la somme des valeurs
arrondies par type si c'est superieur a 0 on ajoute arbitrairement 50
logements a ip et colres,  methodologie locale non officielle.
}
